refactor(web): decompose DailyPlanCard.tsx (1228 → 405 LOC)
Розбиває моноліт DailyPlanCard.tsx на 5 файлів і вирівнює 0013 plan
з реальним станом allowlist-у:

- apps/web/src/modules/nutrition/lib/dailyPlanValidation.ts (139) —
  pure functions (calcMacroKcalMismatch, calcGoalRangeIssues, GOAL_BOUNDS).
- apps/web/src/modules/nutrition/components/DailyPlanWarnings.tsx (209) —
  MacroKcalWarning, MissingMacrosHint, GoalRangeWarning.
- apps/web/src/modules/nutrition/components/DailyPlanMacros.tsx (105) —
  MacroRatioBar, MacroBadge.
- apps/web/src/modules/nutrition/components/DailyPlanMealRow.tsx (127) —
  PlanMeal, MEAL_TYPE_*, DailyPlanMealRow.
- apps/web/src/modules/nutrition/components/DailyPlanGoalSelectors.tsx
  (304) — TDEE/preset dropdowns + outside-click handlers.

DailyPlanCard.tsx тепер 405 effective LOC. Re-exports
calcMacroKcalMismatch/calcGoalRangeIssues для BC з тестами.

eslint.config.js allowlist:
- видалено DailyPlanCard.tsx (тепер 405 LOC, проходить max-lines:600);
- видалено Progress.tsx (579 raw / 546 effective — вже проходить);
- видалено RoutineCalendarPanel.tsx (602 raw / 575 effective — теж).

i18n-allowlist розширено для 4 нових компонентів — рядки тимчасово
лишаються в JSX як були в оригіналі (i18n міграція — окремий issue,
не scope цієї декомпозиції).

0013 doc оновлено: drift reconciliation 2026-05-09, sprint 1 розширено
з 4 до 5 PR-ів (+DailyPlanCard), sprint 2 додано HubDashboard як
drift-пункт, sprint 3 скорочено з 3 до 1 PR-у. FinykApp позначено
як not-needed (537 effective LOC).

Refs: 0013-module-decomposition-round-2.
