+++ apps/web/eslint.i18n-allowlist.json
@@ -153,6 +153,10 @@
   "apps/web/src/modules/nutrition/components/AddMealSheet.tsx",
   "apps/web/src/modules/nutrition/components/BarcodeScanner.tsx",
   "apps/web/src/modules/nutrition/components/DailyPlanCard.tsx",
+  "apps/web/src/modules/nutrition/components/DailyPlanGoalSelectors.tsx",
+  "apps/web/src/modules/nutrition/components/DailyPlanMacros.tsx",
+  "apps/web/src/modules/nutrition/components/DailyPlanMealRow.tsx",
+  "apps/web/src/modules/nutrition/components/DailyPlanWarnings.tsx",
   "apps/web/src/modules/nutrition/components/ItemEditSheet.tsx",
   "apps/web/src/modules/nutrition/components/LogCard.tsx",
   "apps/web/src/modules/nutrition/components/MealRow.tsx",
